Refactored radio_stub send/receive into Connection trait.
