+++ radio/comms-service/src/main.rs
@@ -1,6 +1,6 @@
 extern crate nsl_duplex_d2;
 
 fn main() {
-    let count = nsl_duplex_d2::comms::get_uploaded_file_count().unwrap();
-    println!("File count {}", count);
+    //    let count = nsl_duplex_d2::comms::get_uploaded_file_count().unwrap();
+    //  println!("File count {}", count);
 }
