--- radio/nsl-duplex-d2/tests/get_file.rs
@@ -1,7 +1,0 @@
-extern crate nsl_duplex_d2;
-
-fn main() {
-    let count = nsl_duplex_d2::comms::get_uploaded_file_count().unwrap();
-
-    println!("file count {}", count);
-}
